+++ NEWS.md
@@ -1,7 +1,9 @@
 # bsvarSIGNs 3.0
+
 1. Parallelized code for posterior sampling [#77](https://github.com/bsvars/bsvarSIGNs/issues/77)
 2. Small open economy with fast zero restrictions [#82](https://github.com/bsvars/bsvarSIGNs/issues/82)
 3. Fix forecasting code with Covid hyperparameters [#85](https://github.com/bsvars/bsvarSIGNs/pull/85) [by Fei Shang](https://github.com/lcq110)
+4. Move `spec$prior$estimate_hyper())` to `spec$estimate_hyper()`
 
 1. Updated forecasting formatting [#80](https://github.com/bsvars/bsvarSIGNs/issues/80)